Begin moving experimental noise simulation functionality into main client.
