+++ src/Python/qsharp-core/qsharp/clients/iqsharp.py
@@ -242,28 +242,22 @@ class IQSharpClient(object):
         finally:
             self.display_data_callback = old_callback
 
-    ## Experimental Methods ##
-    # These methods expose experimental functionality that may be removed without
-    # warning. To communicate to users that these are not reliable, we mark
-    # these methods as private, and will re-export them in the
-    # qsharp.experimental submodule.
-
-    def _simulate_noise(self, op, **kwargs) -> Any:
+    def simulate_noise(self, op, **kwargs) -> Any:
         kwargs.setdefault('_timeout_', None)
         return self._execute_callable_magic('experimental.simulate_noise', op, **kwargs)
 
-    def _get_noise_model(self) -> str:
+    def get_noise_model(self) -> str:
         return self._execute(f'%experimental.noise_model')
 
-    def _get_noise_model_by_name(self, name : str) -> None:
+    def get_noise_model_by_name(self, name : str) -> None:
         return self._execute(f'%experimental.noise_model --get-by-name {name}')
 
-    def _set_noise_model(self, json_data : str) -> None:
+    def set_noise_model(self, json_data : str) -> None:
         # We assume json_data is already serialized, so that we skip the support
         # provided by _execute_magic and call directly.
         return self._execute(f'%experimental.noise_model {json_data}')
 
-    def _set_noise_model_by_name(self, name : str) -> None:
+    def set_noise_model_by_name(self, name : str) -> None:
         return self._execute(f'%experimental.noise_model --load-by-name {name}')
